refactor: reorganize project structure and update MySQL schema
