+++ README.md
@@ -6,6 +6,21 @@ A Streamlit-based web application for extracting, analyzing, and storing medical
 
 ---
 
+## 📁 Project Structure
+
+```
+medical_ner_project/
+├── app.py              # Main Streamlit application
+├── config.py           # Configuration and model loading
+├── database.py         # Database operations
+├── run.py              # Convenience run script
+├── requirements.txt    # Python dependencies
+├── .env               # Environment variables (create this)
+└── README.md          # Project documentation
+```
+
+---
+
 ## 🚀 Features
 
 - **PDF Upload:** Upload one or more medical PDF reports.
@@ -70,7 +85,7 @@ CREATE DATABASE medical_ner;
 
 ### Change the NER Model
 
-At the top of `medical_App.py`, set the model you want to use:
+In `config.py`, set the model you want to use:
 
 ```python
 MODEL_NAME = "d4data/biomedical-ner-all"  # Or any other HuggingFace NER model
@@ -82,7 +97,10 @@ You can find more models [here](https://huggingface.co/models?pipeline_tag=token
 
 ## ▶️ Usage
 
-Run the Streamlit app:
+### Run directly with Streamlit
+```bash
+streamlit run app.py
+```
 
 ```bash
 streamlit run medical_App.py
